# contracts/http/v0/openapi.yaml
openapi: 3.1.0
info:
  title: Quivr V2 public text foundation contract
  version: 0.0.0-draft
  description: THE-543 and THE-547 evaluation contracts; endpoint implementations are separate work. Matching
    criterion is plugin-owned and deferred. One configured webhook destination, immutable Matches, independent
    at-least-once Delivery and reference-only notifications. OpenAPI is authoritative for transport shapes.
    THE-640 adds text search and asynchronous Corpus projection rebuild initiation.
security:
- ApiKey: []
paths:
...
  /v0/corpora:
    post:
      operationId: createCorpus
      description: Explicit Corpus creation. Same creation route-family key and canonical request replays
        the same Corpus. Requires corpora:write.
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Corpus'
        default:
          description: Structured error; see contract HTTP mapping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorpusRequest'
      x-required-permissions:
      - corpora:write
    get:
      operationId: listCorpora
      description: Authorized Corpora only. Opaque page cursor bound to action/filter/scope; not a Change
        Cursor or a Record page cursor (either is 422 invalid_cursor).
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorpusPage'
        default:
          description: Structured error; see contract HTTP mapping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: page_cursor
        in: query
        required: false
        schema:
          type: string
          minLength: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 100
      x-required-permissions:
      - corpora:read
...
  /v0/changes:
    get:
      operationId: pollChanges
      description: Same durable journal as SSE. Without cursor, return empty items and current committed
        position as next_cursor (start now). With cursor, return authorized events after it, in commit
        order; duplicates possible. Cursor binds Organization, Corpus filter and authorization scope.
        Expiry is 410 cursor_expired with resync_url, never silent reset; changed scope/filter is 409
        cursor_scope_changed with resync_url. resync_url is the API-relative Record catalog route for
        the Corpus (/v0/records?corpus_id=...). next_cursor advances over scanned events even when none are visible. Public
        events are invalidations; reread current state, do not replay historical content into a current
        mirror.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
        default:
          description: Structured error; see contract HTTP mapping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          minLength: 1
      - name: corpus_id
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 100
      x-required-permissions:
      - changes:read
  /v0/changes/stream:
    get:
      operationId: streamChanges
      description: 'SSE over the same journal. Last-Event-ID takes precedence over query cursor on reconnect.
        id is an opaque Change Cursor; data.event_id is deduplication identity. Named change events carry
        ChangeEvent JSON. checkpoint events carry a cursor when no visible event is emitted, including
        initial start-now checkpoint. Resume cursor expires or scope changes: before headers use HTTP
        410/409; after headers send stream_error with Error JSON then close, without advancing id. Use
        authenticated streaming fetch/http client. See contract for checkpoint framing and resync.'
      responses:
        '200':
          description: Successful response
          content:
            text/event-stream:
              schema:
                type: string
              example: 'id: opaque-cursor

                event: change

                data: {"event_id":"event_1","type":"record.searchable","schema_version":"1","occurred_at":"2026-09-14T12:00:00Z","resource":{"kind":"record","id":"record_1","corpus_id":"corpus_1"},"cursor":"opaque-cursor"}


                '
        default:
          description: Structured error; see contract HTTP mapping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          minLength: 1
      - name: corpus_id
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          type: string
          minLength: 1
      x-required-permissions:
      - changes:read
...
  /v0/search:
    post:
      operationId: searchRecords
      description: Resolve the requested profile, compile mandatory Corpus/Organization prefilters and
        any requested filter, obtain candidates, then canonically hydrate and reauthorize every returned segment. Lexical-first records
        remain eligible without embeddings; semantic-only queries require vector coverage. Profile selection
        does not change access/currentness rules. When a retrieval plugin is pinned, it ranks. It asks
        the engine for candidates in up to three rounds and returns its ranking, which may hold only
        candidates the engine served in this search, each already authorized and hydrated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        default:
          description: Structured error; 400 malformed, 401 unauthenticated, 403 unauthorized scope/action,
            404 absent/inaccessible, 409 idempotency conflict, 422 unsupported_profile, query_too_long
            (the query is over the profile's or the vector space owner's length limit; the message names
            it), unsupported_search or source_filter_unavailable, 502 retrieval_plugin_invalid (the retrieval plugin broke its
            contract, for example ranked a segment the engine never served it), 503 dependency unavailable,
            504 search_deadline_exceeded (the search outran its profile's max_latency_ms).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-required-permissions:
      - content:read
      - search:query
  /v0/search/profiles:
    get:
      operationId: listSearchProfiles
      description: The search profiles this deployment answers, default first. Without a retrieval
        plugin only the built-in default exists; with one, its declared profiles and budgets.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchProfileList'
        default:
          description: Structured error; 401 unauthenticated, 403 unauthorized scope/action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-required-permissions:
      - search:query
  /v0/corpora/{corpus_id}/vector-spaces:
    get:
      operationId: listVectorSpaces
      description: The vector spaces the Corpus's routed Projection Generation carries, the served one
        first, each with its owner (the engine, or the ingestion plugin that declares it), model, dimensions,
        metric, indexed and query modalities, its role in the generation (served answers search, evaluation
        is indexed and compared but never served) and its coverage, the current segments that hold a
        vector in it. A Corpus built before a space was enabled lists only the spaces it was built with;
        rebuild it (rebuildCorpusProjection) to add the others.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorSpaceList'
        default:
          description: Structured error; 401 unauthenticated, 403 unauthorized scope/action, 404 absent/inaccessible,
            503 dependency unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: corpus_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      x-required-permissions:
      - corpora:read
...
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: API key, not necessarily a JWT. Server derives Organization, permitted actions and
        Corpus scope; every resource access is authorized.
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        retryable:
          type: boolean
        field:
          type: string
          minLength: 1
          description: JSON Pointer (RFC 6901) to the request member that caused a 422, when known (for
            example /config/url or /credential/secret/token on connector commands).
        resync_url:
          type: string
          format: uri-reference
      required:
      - code
      - message
      - retryable
...
    RetrievalConfig:
      type: object
      additionalProperties: false
      properties:
        plugin_profile:
          type: string
          minLength: 1
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldMapping'
      required: []
      description: Pin a plugin-provided profile when resolving config. Explicit fields override default
        fields by logical name; unmapped source data remains preserved. getCorpus returns the effective
        resolved fields. The only built-in profile, example.editorial, is illustrative (paired with the example
        extension namespace), not a product default; an uninstalled profile is 422 unsupported_profile.
...
    CorpusRequest:
      type: object
      additionalProperties: false
      properties:
        idempotency_key:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        retrieval:
          $ref: '#/components/schemas/RetrievalConfig'
      required:
      - idempotency_key
      - name
    Corpus:
      type: object
      additionalProperties: false
      properties:
        corpus_id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        effective_retrieval:
          $ref: '#/components/schemas/RetrievalConfig'
      required:
      - corpus_id
      - name
      - effective_retrieval
    VectorSpaceList:
      type: object
      additionalProperties: false
      properties:
        projection_generation_id:
          type: string
          minLength: 1
        segments:
          type: integer
          minimum: 0
          description: Current segments the generation projects; a space whose coverage equals it holds
            a vector for every one.
        items:
          type: array
          items:
            $ref: '#/components/schemas/VectorSpace'
      required:
      - projection_generation_id
      - segments
      - items
...
    CorpusPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Corpus'
        next_page_cursor:
          type: string
          minLength: 1
      required:
      - items
...
    ResourceReference:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          minLength: 1
        id:
          type: string
          minLength: 1
        corpus_id:
          type: string
          minLength: 1
      required:
      - kind
      - id
    ChangeEvent:
      type: object
      additionalProperties: false
      properties:
        event_id:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        schema_version:
          type: string
          minLength: 1
        occurred_at:
          type: string
          format: date-time
        resource:
          $ref: '#/components/schemas/ResourceReference'
        cursor:
          type: string
          minLength: 1
        monitoring:
          $ref: '#/components/schemas/MonitoringReferences'
      required:
      - event_id
      - type
      - schema_version
      - occurred_at
      - resource
      - cursor
      description: Every change to a Record catalog entry emits an event with resource.kind=record and
        resource.id=the affected Record ID. Additional resource-specific events do not replace this invalidation.
        Consumers reread current state; payload detail belongs to THE-547. Monitoring notice types mirror
        WebhookEvent and include monitoring references; event_id identifies that same committed notice.
        Delivery status changes emit delivery.updated events only to the feed, never recursive webhooks.
    ChangePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChangeEvent'
        next_cursor:
          type: string
          minLength: 1
        has_more:
          type: boolean
      required:
      - items
      - next_cursor
      - has_more
...
    SearchRequest:
      type: object
      additionalProperties: false
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 8192
          description: At most 8192 code points on the wire. A semantic or hybrid query is also limited
            by the owner of the searched vector space (the first-party core.ingest plugin accepts at
            most 256 tokens of its model's tokenizer); a longer query is refused with 422
            query_too_long, whose message names the limit, never truncated.
        corpus_ids:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
          maxItems: 16
          uniqueItems: true
        mode:
          type: string
          enum:
          - lexical
          - semantic
          - hybrid
          default: hybrid
        profile:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,31}$
          default: default
          description: A search profile this deployment answers (listSearchProfiles). The built-in
            path answers default only; a pinned retrieval plugin answers the profiles it declares,
            default among them. balanced is a deprecated alias of default, accepted through engine
            0.1.x and removed in engine 0.2.0. An unknown profile returns 422 unsupported_profile.
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
        filter:
          $ref: '#/components/schemas/SearchFilter'
      required:
      - query
      - corpus_ids
      description: Text-only top-k query. Resolve all Corpora in the authenticated Organization and require
        read/search permission for every requested Corpus before querying. Never silently drop an unauthorized
        Corpus. Unknown/unsupported profile or mode returns 422; a dependency outage is an error, not
        an empty successful result. Query token limits are checked against the resolved profile; no silent
        truncation. An optional filter narrows candidates inside the engine query, before ranking and
        the limit, in every mode. Other metadata filters and pagination are outside this surface.
    SearchFilter:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        source_namespaces:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 200
          minItems: 1
          maxItems: 50
          uniqueItems: true
          description: Keep only Records whose Source Namespace is one of these values. Ranking
            and the limit apply within the filtered set, so a source's best matches are returned
            even when other sources outrank them.
      description: Candidate filter applied before ranking. Every present condition must hold.
        A requested Corpus served by a Projection Generation built before source filtering existed
        returns 422 source_filter_unavailable; rebuild that Corpus once (rebuildCorpusProjection)
        to enable it. Unfiltered search is unaffected.
    SearchProfile:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
      required:
      - name
      - version
      description: Resolved retrieval profile identity. Name is the profile that answered (default when
        the request named none or the deprecated balanced). Version identifies what ranked; the built-in
        path's immutable profile version, or plugin:<plugin id>@<version>/<profile> for a retrieval plugin.
    SearchProfileList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchProfileDescription'
          minItems: 1
          maxItems: 8
      required:
      - items
...
    SearchUsage:
      type: object
      additionalProperties: false
      properties:
        rounds:
          type: integer
          minimum: 1
          maximum: 3
        elapsed_ms:
          type: integer
          minimum: 0
        paid_calls:
          type: integer
          minimum: 0
        cost_cents:
          type: number
          minimum: 0
      required:
      - rounds
      - elapsed_ms
      - paid_calls
      - cost_cents
      description: What a search answered by a retrieval plugin spent; rounds of the plugin, elapsed
        time, and the paid calls and cost the plugin reported.
...
    SearchResponse:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
          maxItems: 50
        retrieval_profile:
          $ref: '#/components/schemas/SearchProfile'
        usage:
          $ref: '#/components/schemas/SearchUsage'
      required:
      - items
      - retrieval_profile
      description: Bounded top-k results after canonical rechecks. May contain fewer hits than requested;
        no total count, completeness promise or stable pagination snapshot. Empty results still name the
        resolved profile.
...